remove taskpanel
